+++ blabpy/seedlings/cha.py
@@ -14,8 +14,8 @@ class Parser:
         self.input_file = input_path
         # output == None means we're being called from parse_clan2 main
         if output is None:
-            self.output_file = input_path.replace(".cha", "_processed.csv")
-            self.error_file = input_path.replace(".cha", "_errors.txt")
+            self.output_file = Path(str(input_path).replace(".cha", "_processed.csv"))
+            self.error_file = Path(str(input_path).replace(".cha", "_errors.txt"))
         # otherwise, Parser object is being constructed from an external source
         else:
             processed_path = os.path.split(input_path)[1].replace(".cha", "_processed.csv")
@@ -117,7 +117,7 @@ class Parser:
         personal_info_com_start = None
         personal_info_com_end = None
 
-        with open(self.input_file, "rU") as input:
+        with open(self.input_file, "r") as input:
             for index, line in enumerate(input):
 
                 if line.startswith("%com:") and ("|" not in line):
@@ -649,7 +649,7 @@ class Parser:
         :return: returns false if we found a problem
         """
 
-        with open(self.input_file, "rU") as input:
+        with open(self.input_file, "r") as input:
             for index, line in enumerate(input):
                 regx_result = self.interval_regx.findall(line)
                 if regx_result: